+++ routes/create.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+
+module.exports = (db) => {
+  router.get("/", (req, res) => {
+    res.render("index");
+  });
+
+  return router;
+};
